Issue #1138: Switched to a config Builder Helper
